requirements copy.txt
about-time==4.2.1
alive-progress==3.2.0
azure-core==1.32.0
azure-cosmos==4.9.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
et_xmlfile==2.0.0
grapheme==0.6.0
idna==3.10
mypy-extensions==1.0.0
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
